requirements.txt
Django>=4.0,<5.0
channels>=4.0,<5.0
daphne>=4.0,<5.0
channels_redis>=4.0,<5.0
kafka-python>=2.0,<3.0